feature: Lego.
Lego is the ability to compose Svalboard firmwares like
legos.  Ballpoint and pointball disappear.  Because they
are just a trackball firmware and a trackpoint firmware,
flashed on the correct sides.

If there is a blank, it can not be master.

Trackpoints must be master.

Trackballs don't care, except that they must be master
if there is nothing else.